+++ Session07/src/bai2/Main.java
@@ -8,18 +8,18 @@ public class Main {
 
         System.out.println("Đơn hàng: tổng tiền 1.000.000, áp dụng PercentageDiscount 10%");
         OrderCalculator calc1 = new OrderCalculator(new PercentageDiscount(10));
-        System.out.println("Số tiền sau giảm: " + (long)calc1.calculate(total));
+        System.out.println("Số tiền sau giảm: " + calc1.calculate(total));
 
         System.out.println("\nĐơn hàng: tổng tiền 1.000.000, áp dụng FixedDiscount 50.000");
         OrderCalculator calc2 = new OrderCalculator(new FixedDiscount(50000));
-        System.out.println("Số tiền sau giảm: " + (long)calc2.calculate(total));
+        System.out.println("Số tiền sau giảm: " + calc2.calculate(total));
 
         System.out.println("\nĐơn hàng: tổng tiền 1.000.000, áp dụng NoDiscount");
         OrderCalculator calc3 = new OrderCalculator(new NoDiscount());
-        System.out.println("Số tiền sau giảm: " + (long)calc3.calculate(total));
+        System.out.println("Số tiền sau giảm: " + calc3.calculate(total));
 
         System.out.println("\nThêm HolidayDiscount 15%");
         OrderCalculator calc4 = new OrderCalculator(new HolidayDiscount());
-        System.out.println("Số tiền sau giảm: " + (long)calc4.calculate(total));
+        System.out.println("Số tiền sau giảm: " + calc4.calculate(total));
     }
 }
